+++ src/regular_polygon2d.h
@@ -7,7 +7,7 @@ using namespace godot;
 class RegularPolygon2D : public Polygon2D {
     GDCLASS(RegularPolygon2D, Polygon2D);
 
-    int nsides;
+    long nsides;
     double radius;
 
 protected:
@@ -15,8 +15,8 @@ protected:
 
 public:
 
-    void set_nsides(int nsides_p);
-    size_t get_nsides();
+    void set_nsides(long nsides_p);
+    int get_nsides();
 
     void set_radius(double radius_p);
     double get_radius();
